+++ pkg/parsergen/backend/typescript/doc.go
@@ -1,0 +1,3 @@
+// Package typescript provides a parser generator backend which takes a parse table as input and outputs TypeScript
+// source code.
+package typescript