+++ pkg/scannergen/backend/typescript/doc.go
@@ -1,0 +1,3 @@
+// Package typescript provides a scanner generator backend which takes a DFA as input and outputs TypeScript source
+// code.
+package typescript